third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/full-block-post-from-iframe-expected.txt

CONSOLE ERROR: line 4: The XSS Auditor blocked access to 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl' because the source code of a script was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.


--------
Frame: 'frame'
--------
Could not load the requested resource.
Error code: -28 (net::ERR_BLOCKED_BY_XSS_AUDITOR)
